+++ index.html
@@ -235,7 +235,7 @@
     </template>
 
   <script type="text/javascript" src="js/functions.js"></script>
-  <script type="text/javascript" src="js/main.js"></script>
+  <script type="module" src="js/main.js"></script>
 
 </body>
 </html>
